# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Central Auth System
  version: 1.0.0
  description: API for central authentication

paths:
  /auth/signup:
    post:
      summary: Create user
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                firstName:
                  type: string
                lastName:
                  type: string

      responses:
        '201':
          description: create user and send verification email
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'User created. Please check your email to verify your account'
                  data:
                    type: object
                    required:
                      - id
                      - email
                      - firstName
                      - lastName
                      - status
                      - emailVerified
                    properties:
                      id:
                        type: string
                        format: uuid
                        example: '123e4567-e89b-12d3-a456-426614174000'
                      email:
                        type: string
                        format: email
                        example: 'user@example.com'
                      status:
                        type: string
                        enum:
                          [
                            PENDING,
                            ACTIVE,
                            SUSPENDED,
                            LOCKED,
                            DEACTIVATED,
                            DELETED,
                          ]
                        example: 'PENDING'
                      firstName:
                        type: string
                        example: 'John'
                      lastName:
                        type: string
                        example: 'Doe'
                      emailVerified:
                        type: boolean
                        example: true
        # Error responses (400 bad request, 409 conflict, ...)
        '400':
          description: Bad request (prisma errors)
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: 'Invalie credentials'

        '409':
          description: Conflict exception(User with email already exists)
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: 'User with this email already exists'
        '500':
          description: Internal server error

  /auth/verify-email:
    get:
      summary: Verify email
      operationId: verifyEmail
      parameters:
        - in: query
          name: token
          required: true
          description: Verification token sent to the user's email
          schema:
            type: string
            example: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'email verified successfully. You can log in'
        '400':
          description: Bat request (Token not found, Invalid or used token)
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 400
                  message:
                    type: string
                    example: 'Invalid or used token'

  /auth/login:
    post:
      summary: Login user
      security:
        - bearerToken: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: login user
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - refreshToken
                  - tokenType
                  - expiresIn
                properties:
                  accessToken:
                    type: string
                    example: E3ODQ3MTYxMjksImV4cCI6MTc4NDcxNzAyOX0...
                  refreshToken:
                    type: string
                    example: 9uIjowLCJpYXQiOjE3ODQ3MTYxMjksImV4cCI6MTc4NzMwODEyOX0...
                  tokenType:
                    type: string
                    example: bearer
                  expiresIn:
                    type: number
                    example: 900

  /auth/refresh:
    post:
      summary: refresh token
      security:
        - bearerToken: []
      operationId: refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  format: token
      responses:
        '200':
          description: refresh token
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - refreshToken
                  - tokenType
                  - expiresIn
                properties:
                  accessToken:
                    type: string
                    example: E3ODQ3MTYxMjksImV4cCI6MTc4NDcxNzAyOX0...
                  refreshToken:
                    type: string
                    example: 9uIjowLCJpYXQiOjE3ODQ3MTYxMjksImV4cCI6MTc4NzMwODEyOX0...
                  tokenType:
                    type: string
                    example: bearer
                  expiresIn:
                    type: number
                    example: 900F

  /auth/logout:
    post:
      summary: logout user
      security:
        - bearerToken: []
      responses:
        '200':
          description: logout user
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: user logged out successfully
                  data:
                    type: object
                    nullable: true
                    example: null

  /auth/sessions/{id}/revoke:
    post:
      summary: revoke token
      security:
        - bearerToken: []
      operationId: revokeSession
      parameters:
        - in: path
          name: id
          required: true
          description: Session ID to revoke
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: session revoked successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                  - data
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: session revoked successfully
                  data:
                    type: object
                    nullable: true
                    example: null
        '401':
          description: User not authenticated or session not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: Session not found

  /auth/sessions:
    get:
      summary: get sessions
      security:
        - bearerToken: []
      operationId: getUserSessions
      responses:
        '200':
          description: get active sessions for current user
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required:
                    - id
                    - userId
                    - lastActivity
                    - expiresAt
                    - idleExpiresAt
                    - refreshVersion
                    - createdAt
                  properties:
                    id:
                      type: string
                      format: uuid
                    userId:
                      type: string
                    ip:
                      type: string
                      nullable: true
                    userAgent:
                      type: string
                      nullable: true
                    lastActivity:
                      type: string
                      format: date-time
                    expiresAt:
                      type: string
                      format: date-time
                    idleExpiresAt:
                      type: string
                      format: date-time
                    revokedAt:
                      type: string
                      format: date-time
                      nullable: true
                    revokedReason:
                      type: string
                      nullable: true
                    refreshVersion:
                      type: integer
                    createdAt:
                      type: string
                      format: date-time
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: User not authenticated

  /auth/password/forgot:
    post:
      summary: reset password
      operationId: reset password
      retuestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
      responses:
        
  /organizations:
    post:
      summary: create organization
      security:
        - bearerToken: []
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - slug
                - taxNo
              properties:
                name:
                  type: string
                  example: Acme Inc
                slug:
                  type: string
                  example: acme-inc
                taxNo:
                  type: string
                  example: 123456789
      responses:
        '201':
          description: organization created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - name
                  - slug
                  - createdAt
                properties:
                  id:
                    type: string
                    example: 01K1E7G8VGVQ7KV9ZQV2XH2EAW
                  name:
                    type: string
                    example: Acme Inc
                  slug:
                    type: string
                    example: acme-inc
                  taxNo:
                    type: string
                    nullable: true
                    example: 123456789
                  address:
                    type: string
                    nullable: true
                    example: null
                  createdAt:
                    type: string
                    format: date-time
        '401':
          description: User not authenticated
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 401
                  message:
                    type: string
                    example: User not authenticated
        '409':
          description: Organization name or slug already exists
          content:
            application/json:
              schema:
                type: object
                required:
                  - statusCode
                  - message
                properties:
                  statusCode:
                    type: integer
                    example: 409
                  message:
                    type: string
                    example: Organization name or slug already exists

  /organizations/{id}/invitations:
    post:
      summary: invite user to organization
      security:
        - bearerToken: []
      operationId: inviteOrganizationMember
      parameters:
        - in: path
          name: id
          required: true
          description: Organization ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - role
              properties:
                email:
                  type: string
                  format: email
                  example: member@acme.com
                role:
                  type: string
                  enum: [OWNER, ADMIN, MANAGER, DIRECTOR, EMPLOYEE, VIEWER]
                  example: EMPLOYEE
                expiresAt:
                  type: string
                  format: date-time
                  nullable: true
                  example: 2026-08-31T23:59:59Z
      responses:
        '201':
          description: invitation created
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - organizationId
                  - email
                  - roleId
                  - token
                  - expiresAt
                properties:
                  id:
                    type: string
                    format: uuid
                  organizationId:
                    type: string
                  email:
                    type: string
                    format: email
                  roleId:
                    type: string
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  acceptedAt:
                    type: string
                    format: date-time
                    nullable: true
        '401':
          description: User not authenticated
        '403':
          description: Not enough permission to invite members
        '404':
          description: Organization or role not found
        '409':
          description: Invitation already exists for this email

  /organizations/invitations/accept:
    post:
      summary: accept invitation
      security:
        - bearerToken: []
      operationId: acceptOrganizationInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  example: 7787f2b5-b711-4fb9-8abf-2f813b7b80b3
      responses:
        '200':
          description: invitation accepted
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: invitation accepted
        '400':
          description: Invalid, expired, or already used invitation token
        '401':
          description: User not authenticated
        '404':
          description: Invitation not found

  /organizations/{id}/members/{memberId}/role:
    patch:
      summary: assign role to user
      security:
        - bearerToken: []
      operationId: assignOrganizationMemberRole
      parameters:
        - in: path
          name: id
          required: true
          description: Organization ID
          schema:
            type: string
        - in: path
          name: memberId
          required: true
          description: Organization member ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - role
              properties:
                role:
                  type: string
                  enum: [OWNER, ADMIN, MANAGER, DIRECTOR, EMPLOYEE, VIEWER]
                  example: MANAGER
      responses:
        '200':
          description: member role updated
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: member role updated
        '401':
          description: User not authenticated
        '403':
          description: Not enough permission to assign roles
        '404':
          description: Organization, member, or role not found

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: 01KXG8848PVGZBKQKZ5S1H5J20
    Organization:
      type: object
    OrganizationMember:
      type: object
    Role:
      type: object
    Permission:
      type: object
    RolePermission:
      type: object
    Invitation:
      type: object
    Session:
      type: object
    UserToken:
      type: object
    AuditLog:
      type: object
    ErrorResponse:
      type: object
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      bearerFormat: JWT